basic cleanup
