+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/QuizServiceApplication.java
@@ -8,7 +8,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
  * This is the entry point of the application.
  */
 @SpringBootApplication
-@EntityScan({"de.unistuttgart.iste.gits.common.resource_markdown", "de.unistuttgart.iste.gits.quiz_service.persistence.dao"})
+@EntityScan({"de.unistuttgart.iste.gits.common.resource_markdown", "de.unistuttgart.iste.gits.quiz_service.persistence.entity"})
 public class QuizServiceApplication {
 
     public static void main(String[] args) {
