+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/controller/SubscriptionController.java
@@ -6,10 +6,7 @@ import io.dapr.Topic;
 import io.dapr.client.domain.CloudEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
 
 import java.util.Map;
@@ -28,7 +25,7 @@ public class SubscriptionController {
     @PostMapping(path = "/quiz-service/content-changes-pubsub")
     public Mono<Void> updateAssociation(@RequestBody CloudEvent<ContentChangeEvent> cloudEvent, @RequestHeader Map<String, String> headers) {
 
-        return Mono.fromRunnable(() -> quizService.removeContentIds(cloudEvent.getData()));
+        return Mono.fromRunnable(() -> quizService.deleteQuizzesWhenQuizContentIsDeleted(cloudEvent.getData()));
     }
 
 }
