+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/persistence/entity/ClozeQuestionEntity.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.gits.quiz_service.persistence.dao;
+package de.unistuttgart.iste.gits.quiz_service.persistence.entity;
 
 
 import jakarta.persistence.*;
