+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/persistence/entity/ExactAnswerQuestionEntity.java
@@ -1,27 +1,32 @@
-package de.unistuttgart.iste.gits.quiz_service.persistence.dao;
+package de.unistuttgart.iste.gits.quiz_service.persistence.entity;
 
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEmbeddable;
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEntity;
-import de.unistuttgart.iste.gits.generated.dto.ClozeElementType;
 import jakarta.persistence.*;
 import lombok.*;
+import lombok.experimental.SuperBuilder;
 
-@Embeddable
-@Data
-@Builder
+import java.util.List;
+
+@Entity(name = "ExactAnswerQuestion")
+@ToString(callSuper = true)
+@Getter
+@Setter
+@EqualsAndHashCode(callSuper = true)
+@SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ClozeElementEmbeddable {
-
-    @Enumerated(EnumType.ORDINAL)
-    private ClozeElementType type;
+public class ExactAnswerQuestionEntity extends QuestionEntity {
 
     @Embedded
     @Builder.Default
     private ResourceMarkdownEmbeddable text = new ResourceMarkdownEmbeddable("");
 
-    @Column(nullable = true)
-    private String correctAnswer;
+    @ElementCollection
+    private List<String> correctAnswers;
+
+    @Column(nullable = false)
+    private boolean caseSensitive;
 
     @OneToOne(optional = true, cascade = CascadeType.ALL)
     private ResourceMarkdownEntity feedback;
