+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/persistence/entity/NumericQuestionEntity.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.gits.quiz_service.persistence.dao;
+package de.unistuttgart.iste.gits.quiz_service.persistence.entity;
 
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEmbeddable;
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEntity;
@@ -6,9 +6,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.util.List;
-
-@Entity(name = "ExactAnswerQuestion")
+@Entity(name = "NumericQuestion")
 @ToString(callSuper = true)
 @Getter
 @Setter
@@ -16,17 +14,17 @@ import java.util.List;
 @SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ExactAnswerQuestionEntity extends QuestionEntity {
+public class NumericQuestionEntity extends QuestionEntity {
 
     @Embedded
     @Builder.Default
     private ResourceMarkdownEmbeddable text = new ResourceMarkdownEmbeddable("");
 
-    @ElementCollection
-    private List<String> correctAnswers;
+    @Column(nullable = false)
+    private double correctAnswer;
 
     @Column(nullable = false)
-    private boolean caseSensitive;
+    private double tolerance;
 
     @OneToOne(optional = true, cascade = CascadeType.ALL)
     private ResourceMarkdownEntity feedback;
