+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/persistence/entity/SelfAssessmentQuestionEntity.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.gits.quiz_service.persistence.dao;
+package de.unistuttgart.iste.gits.quiz_service.persistence.entity;
 
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEmbeddable;
 import de.unistuttgart.iste.gits.common.resource_markdown.ResourceMarkdownEntity;
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-@Entity(name = "NumericQuestion")
+@Entity(name = "SelfAssessmentQuestion")
 @ToString(callSuper = true)
 @Getter
 @Setter
@@ -14,18 +14,13 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor
-public class NumericQuestionEntity extends QuestionEntity {
+public class SelfAssessmentQuestionEntity extends QuestionEntity {
 
     @Embedded
     @Builder.Default
     private ResourceMarkdownEmbeddable text = new ResourceMarkdownEmbeddable("");
 
-    @Column(nullable = false)
-    private double correctAnswer;
-
-    @Column(nullable = false)
-    private double tolerance;
-
     @OneToOne(optional = true, cascade = CascadeType.ALL)
-    private ResourceMarkdownEntity feedback;
+    @Builder.Default
+    private ResourceMarkdownEntity solutionSuggestion = new ResourceMarkdownEntity("");
 }
