+++ src/main/java/de/unistuttgart/iste/gits/quiz_service/persistence/mapper/QuizMapper.java
@@ -1,7 +1,7 @@
 package de.unistuttgart.iste.gits.quiz_service.persistence.mapper;
 
 import de.unistuttgart.iste.gits.generated.dto.*;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.*;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.*;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
@@ -20,13 +20,43 @@ public class QuizMapper {
     public Quiz entityToDto(QuizEntity entity) {
         // manual mapping necessary because of QuestionInterface
         // which cannot automatically be mapped by model mapper
-        return Quiz.builder()
+        Quiz result = Quiz.builder()
                 .setAssessmentId(entity.getAssessmentId())
                 .setQuestionPoolingMode(entity.getQuestionPoolingMode())
                 .setNumberOfRandomlySelectedQuestions(entity.getNumberOfRandomlySelectedQuestions())
                 .setRequiredCorrectAnswers(entity.getRequiredCorrectAnswers())
                 .setQuestionPool(entity.getQuestionPool().stream().map(this::questionEntityToDto).toList())
                 .build();
+        return selectQuestionsInQuiz(result);
+    }
+
+    /**
+     * Selects the questions for a quiz based on the question pooling mode.
+     * <p>
+     * If the question pooling mode is {@link QuestionPoolingMode#ORDERED}, all questions are selected
+     * and the order is preserved.
+     * If the question pooling mode is {@link QuestionPoolingMode#RANDOM}, the questions are shuffled
+     * and the number of questions is limited to {@link Quiz#getNumberOfRandomlySelectedQuestions()}.
+     *
+     * @param quiz the quiz
+     * @return the quiz with the selected questions
+     */
+    private Quiz selectQuestionsInQuiz(Quiz quiz) {
+        if (quiz.getQuestionPoolingMode() == QuestionPoolingMode.ORDERED) {
+            quiz.setSelectedQuestions(quiz.getQuestionPool());
+            return quiz;
+        }
+
+        int limit = quiz.getQuestionPool().size();
+        if (quiz.getNumberOfRandomlySelectedQuestions() != null) {
+            limit = Math.min(limit, quiz.getNumberOfRandomlySelectedQuestions());
+        }
+
+        List<Question> pool = new ArrayList<>(quiz.getQuestionPool());
+        Collections.shuffle(pool);
+        quiz.setSelectedQuestions(pool.subList(0, limit));
+
+        return quiz;
     }
 
     private Question questionEntityToDto(QuestionEntity questionEntity) {
