+++ src/test/java/de/unistuttgart/iste/gits/quiz_service/api/mutation/MutateQuizAddMultipleChoiceQuestionTest.java
@@ -5,7 +5,7 @@ import de.unistuttgart.iste.gits.common.testutil.TablesToDelete;
 import de.unistuttgart.iste.gits.generated.dto.*;
 import de.unistuttgart.iste.gits.quiz_service.TestData;
 import de.unistuttgart.iste.gits.quiz_service.api.QuizFragments;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.QuizEntity;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.QuizEntity;
 import de.unistuttgart.iste.gits.quiz_service.persistence.repository.QuizRepository;
 import graphql.ErrorType;
 import org.junit.jupiter.api.Test;
