+++ src/test/java/de/unistuttgart/iste/gits/quiz_service/api/mutation/MutateQuizAddSelfAssessmentQuestionTest.java
@@ -6,7 +6,7 @@ import de.unistuttgart.iste.gits.generated.dto.CreateSelfAssessmentQuestionInput
 import de.unistuttgart.iste.gits.generated.dto.ResourceMarkdownInput;
 import de.unistuttgart.iste.gits.quiz_service.TestData;
 import de.unistuttgart.iste.gits.quiz_service.api.QuizFragments;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.*;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.*;
 import de.unistuttgart.iste.gits.quiz_service.persistence.repository.QuizRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Test;
