+++ src/test/java/de/unistuttgart/iste/gits/quiz_service/api/mutation/MutateQuizUpdateSelfAssessmentQuestionTest.java
@@ -6,8 +6,8 @@ import de.unistuttgart.iste.gits.generated.dto.ResourceMarkdownInput;
 import de.unistuttgart.iste.gits.generated.dto.UpdateSelfAssessmentQuestionInput;
 import de.unistuttgart.iste.gits.quiz_service.TestData;
 import de.unistuttgart.iste.gits.quiz_service.api.QuizFragments;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.QuizEntity;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.SelfAssessmentQuestionEntity;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.QuizEntity;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.SelfAssessmentQuestionEntity;
 import de.unistuttgart.iste.gits.quiz_service.persistence.repository.QuizRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Test;
