+++ src/test/java/de/unistuttgart/iste/gits/quiz_service/matcher/QuizEntityToCreateInputMatcher.java
@@ -1,7 +1,7 @@
 package de.unistuttgart.iste.gits.quiz_service.matcher;
 
 import de.unistuttgart.iste.gits.generated.dto.CreateQuizInput;
-import de.unistuttgart.iste.gits.quiz_service.persistence.dao.QuizEntity;
+import de.unistuttgart.iste.gits.quiz_service.persistence.entity.QuizEntity;
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeDiagnosingMatcher;
